Used more appropriate data types
* The specification stated a non-negative integer so switching from `int` to `uint` means we have to deal with larger values
* We know we can only return a maximum value of 9 so we don't need the storage of an `int`

+++ src/AddDigits/DigitsAdder.cs
@@ -2,16 +2,16 @@
 
 public static class DigitsAdder
 {
-    public static int AddDigits(int value)
+    public static byte AddDigits(uint value)
     {
         var sum = value
             .ToString()
             .ToCharArray()
-            .Select(c => (int)char.GetNumericValue(c)).Sum(x => x);
+            .Select(c => (uint)char.GetNumericValue(c)).Sum(x => x);
 
         // Tail recursion... Will this be inefficient on very large numbers? Unlikely as with int we can't get up
-        // to three digits on the first iteration. The highest value we can get is 82. This limits the sum. I _think_
+        // to three digits on the first iteration. The highest value we can get is 84. This limits the sum. I _think_
         // we will need at most three iterations, so extra stack frames is probably not a concern to us
-        return sum >= 10 ? AddDigits(sum) : sum;
+        return sum >= 10 ? AddDigits((uint)sum) : (byte)sum;
     }
 }
